feat: Add initialJobs prop to JobList component
The `JobList` component now accepts an `initialJobs` prop, which allows passing in an initial array of job postings. This prop is used to set the initial state of the `jobs` state variable. If the `initialJobs` array is empty, the component will fetch jobs from the server. This change improves the flexibility and reusability of the `JobList` component.